Move runOptions stream checks before the `for` loop to prevent cpu load
`runOptions.stream` is now `true` by default, so `Listen` now performs
a lot of empty runs, causing cpu load (one core having 100% load all the time).

This fix prevents the overload by never running the loop if `stream` is set
to `true`, as all cases in the loop relied on it being false.