fix(yarn-pnp): fixing generateBinPath when yarn pnp is enabled
The call to require.resolve(turbo) was failing because the package does not have a main export
defined in the package.json, nor is there an index.js. Since we are just looking for the directory
the package is unplugged into, we can resolve the path to the package.json and get the dirname
of that.