Generate JSON Schema from Golang structs
This commit continues the amazing work done in #634 and #726.

Instead of generating the JSON schema from a separate TypeScript file,
this commit introduces a new CLI (to avoid bloating the main one) that
will output a JSON schema from the struct definition itself, including
the actual code comments.

This will allow to better integration and to avoid differences between
the two codebases.