fix(dep-graph): allow mix of internal and external deps with same name (#802)
This is my first time coding with golang, so please let me know if there is anything non-idiomatic or where things can be written better. 

When turbo currently tries to build the dependency graph, it doesn't consider the specified
version/range of the dependency, it just assumes that if that package name exists in the
workspace it must be an internal dependency. Because of this, cycles are detected and
the process hangs.

This PR adds some additional logic to ensure that dependencies are only considered
internal if the local/workspace package's version matches the semver range specified for
the dependency.

Fixes: #796