Extract ResolvePackages (#830)
 * Move package resolution logic to `scope` package
 * Remove unused functionality from `SCM` interface to make it easier to mock in tests
 * Start adding tests for scope resolution
 
**Review Note:**
`scope_test.go` provides examples of parsed flags and how they translate to packages considered in-scope. I've done my best to infer what packages should be built in each scenario, but please verify. Also very happy to accept additions to that test.

**Possibly controversial change:**
To maintain backwards compatibility, specifying `--since` and `--scope` to `turbo run` will imply `--include-dependencies`. This means that using scope will generally cause all dependencies of the specified scopes to be built, although the cache will still be used so this may potentially be cheap. 

Previously, this behavior was triggered when `--scope` and `--since` constraints produced packages to be built. In cases where they were both specified but one or both did not produce packages to be built, dependencies were not included. 

With this change, the behavior will be consistent based on commandline flags and default to building dependencies. With an upcoming switch over to `cobra` for commandline parsing, we should be able to flag off building dependencies if the user requests it.