Allow specifying cache outputs outside of package folder (#861)
This allows people to specify `outputs` outside of the package folder from which the command originated, which is especially useful for code-generation tools and polyglot repos. To enable this, we store artifacts in the cache relative to the root of the monorepo (see screenshot) below for before (top) and after (bottom).

<img width="419" alt="CleanShot 2022-03-10 at 12 58 21@2x" src="https://user-images.githubusercontent.com/4060187/157725834-a19ffb57-d310-41d7-b44e-b95ddcd48a17.png">

Since we are changing how cache is structured, I updated the global cache key in turbo in order to prevent prior old artifacts from ever ever being cache hits with after this change

The only downside to this change, which might not be a downside at all, I guess, is possibility that artifacts are now tied to the directory structure of outputs relative to the root of the monorepo.