fix(remote-cache): use posix paths for tar archives (#895)
This should ensure that archives generated for remote caching have the directory
structure preserved in a consistant manner to what is generated on linux/mac

Fixes: #801

PR #891 I believe fixes the hash calculation so that a remote-cache generated on linux or mac can be consumed on windows and no longer results in a cache miss, but I still have concerns about consuming archives generated on windows from the linux environment. In particular, these paths are coming from globby and don't pass through the normalization from the linked PR, so they are still os-dependent.

I have encountered some problems with builds failing in GH Actions because dependencies that should have been built already were not found when I had previously built those projects on my windows machine. I believe this PR should fix the problem but I haven't had the chance to try reproducing the existing issue to verify that this fixes it. I might be able to get that done later today, but I wanted to get a PR up sooner rather than later