Only check for yarn berry if it is installed (#976)
Fix https://github.com/vercel/turborepo/issues/962

In a follow up, we ideally should not need to make an `exec` call to start `yarn` which is written in node.js just to determine it's version. All we need to know is if it is greater than 2. We Also don't need to be parsing full semver. As soon as we know the first character is a >= 2 we know it is berry