feat(colors): add options for forcing color/no-color
This change checks for --color / --no-color as command-line arguments as a mechanism for forcing color in non-TTY situations, or forcing --no-color for TTY. Additionally, this change hooks into the FORCED_COLOR environment variable that is used by the [`supports-color` npm package](https://www.npmjs.com/package/supports-color) that is used by a variety of packages in the node ecosystem.

The intention of this PR is to alleviate some of the concerns from Issue #897 by allowing users of turborepo to enable color output on CI runners such as Github Actions which support color output (as mentioned in the [fatih/color documentation](https://github.com/fatih/color/tree/master#github-actions)). With github actions and local builds are both running in color, then the cached output should be consistently in color, compared to the current situation where its a mix of color and non-color output with a remote shared between CI and users building interactively.

The FORCED_COLOR environment variable is something that I find myself already setting when running turbo so that the streamed output is in color
for the build and testing tools that I am running with it, so it seemed like a
natural way to provide this argument without needing to modify the CLI parameters being passed to turbo.

In the event multiple methods are being used, the precedence is:
* isTTY or not (provided by fatih/color)
* $NO_COLOR (if set, then color is disabled)
* $FORCED_COLOR ("1", "2", "3", "true" will enable color, "0", or "false" will disable color)
* --color/--no-color (the last argument wins if both exist)