Add back support for `link:...` dependencies
This is similar to the work done for `file:...` dependencies in
32cebb1094bfb572fcdff01c0cb6d09f7538b418.

Check if file paths pointed to by `link:...` are within the monorepo.
If so, consider them internal dependencies, and otherwise consider them as external dependencies.

This can be useful because pnpm uses the `link:` protocol when using
`pnpm install ./package-a/`.