Implement globby.GlobFiles using doublestar.GlobWalk (#1108)
This change aims to improve the performance of our file walking by only walking the portions of the directory that we care about. In this iteration it aims to be 100% backwards compatible with the previous implementation, excepting one obvious bug, resulting in a boolean change to the tests.

It also adopts a set instead of a range for storage as doublestar will happily walk a directory multiple times if you feed it a nonsensical glob pattern such as `**/**/**`.