Refactor turbo_json (#1152)
In preparing to hide some package manager details I discovered this code was unnecessarily coupled to `package_json`, and in multiple places in the codebase.

Split commits to make it easier to follow:
1. Nothing more than splitting the `package_json` into two pieces.
2. Move code from `config.go` to its new home.
3. Rename symbols away from `turboConfig` which has a confusable in `TurborepoConfig` from the login work.

A pretty-obvious bit of followup work is to move both `package_json` and `turbo_json` out of the `fs` package—they could be trivially refactored to never touch the file system (and possibly should be for testability reasons).