Better logging for failing to upload an artifact (#1201)
As mentioned here https://github.com/vercel/turborepo/issues/487, in some situations I'd prefer to know which artifact is so large so I can track it down (most of our artifacts should not be so large).

Also if there's a place up the stacktrace I can insert the package name, that'd be really ideal, but that looks like a bigger lift so want to make sure we want to do that.