Adopt doublestar, disable traversing symlinks in globwalk (#1275)
Commits:
 * Copy in `doublestar` files, point our existing code at them
 * Make them satisfy our linter
 * Remove symlink traversal and make associated test changes

Fixes #1265 

Our globwalk when collecting outputs was traversing symlinks. This caused us to copy their pointed-to contents as if the symlink was a directory, and then when copying the link itself, throw an error because it already existed (as a directory...). This in turn would crash all of our file-copying worker goroutines, leaving the main goroutine waiting indefinitely for file copying to complete.

Not included here, but we should potentially make our file-copying goroutines more robust so that at the very least we get a better error message.