Create noopCache to ensure that users always have a valid cache object. (#1317)
This makes it possible for a repository whose CI configuration is `--remote-only` to work when receiving pull-requests from forked repositories that do not have access to secrets configuration in CI.

Example failures:
- https://github.com/vercel/turborepo/pull/1302: [Error](https://github.com/vercel/turborepo/runs/6681959655?check_suite_focus=true#step:8:56)
- https://github.com/vercel/vercel/pull/7575: [Error](https://github.com/vercel/vercel/runs/6599439965?check_suite_focus=true#step:10:12)

One example workaround:
https://github.com/vercel/vercel/pull/7873

However, since this does not prevent the build from running to completion, I do not believe that it should be a hard error. While in the process of addressing this I also discovered that, since the `httpCache` can remove itself, there is an alternative path to triggering a panic if the `httpCache` is the only cache you have enabled and it removes itself.

Using a `noopCache` as a cache of last resort (similar to the [upstream solution](https://github.com/thought-machine/please/commit/e0b5e2900f5caab98a20d476d34bce66da28cdb8)) resolves both of these issues.

Related to: #1169