Port prune command to cobra (#1300)
Ports `prune` over to `cobra`. There is still a fair amount to be done to refactor prune, but this is a start that allows easier expansion of commandline configuration.

Fixes #1258

Adds the feature requested in #1159 (flag for outputs of `prune`)

Note that this PR stacks on top of porting over `run`