Prevent side effects when looking for config files. (#1328)
As noted in https://github.com/vercel/turborepo/pull/1156#discussion_r863341102, `xdg.ConfigFile` has side effects. One of the consequences of the side effects is that sometimes, when simply wanting to check for existence, we would error out entirely because the path wasn't writable.

This splits the process of creating and reading a config file path, which means that the error won't be encountered unless a user is explicitly trying to create a configuration file. (Which *should* be an error.)

Fixes #1270