Log to temp files then rename (#1351)
Log files are hard-linked into the cache, so when re-running a task we need to ensure we don't pollute a previous run. 

This changes our logging to stream to a temp file, which we then rename to the actual log file. This serves to break the hard link. 

Note that:
 * renames will succeed even if the target already exists, as long as the target isn't a directory. All of our targets are log files
 * renames can fail if the rename is across filesystem boundaries. Our renames are within the same directory.
 
 Fixes #1323