Update schema generation script invocation. (#1436)
This is kind of a weird fix. I believe it works by resetting the execution root for the `compileProgram` execution; not based upon any actual configuration in our TSConfig file.

The previous state would succeed or fail based upon the `node_modules` directory structure, this changes the execution location to be from a known-good location.