Fix tailwind example (#1500)
The wrong npm package was installed in the tailwind example `tailwind` => `tailwindcss`.

I also added exports field for styles.css as it was not found from within the `_app.tsx` in next applications, but I am not sure if this is required.

```
web:dev: error - ./pages/_app.tsx:3:0
web:dev: Module not found: Package path ./styles.css is not exported from package
~/turborepo/examples/with-tailwind/node_modules/ui (see exports field in ~/turborepo/examples/with-tailwind/node_modules/ui/package.json)
```