Create more robust arg passing testing script
The initial testing script had a hidden dependency on the length of the node path which would trigger the output to be spread over multiple lines.
This lead to confusing logic around handling the case when it all fit onto one line.
This commit changes the script to now always output all information on a single line.