output to terminal when we are just waiting for cache uploads (#1686)
As a stopgap before we move cache uploads to the daemon, I added some terminal output if uploading is blocking process completion. 

While obviously not ideal that users are just waiting for cache population, I think it's at least better that they know that that is what is happening, rather than the process appearing to hang.