Move parent_path into FileSystemPathVc (#265)
This moves `FileSystem::parent_path` (which takes a `FileSystemPathVc`) to be directly on `FileSystemPathVc`. There's already a `FileSystemPathVc::parent` method which internally calls `FileSystem::parent_path` with self, which seems to be an indirect way to implement this.

With the new method location, we're able to take advantage of `FileSystemPathVc::new_normalized` to skip some unnecessary work. And we also reduce duplication, there were two nearly identical implementations of `parent_path` between `DevServerFileSystem` and `DiskFileSystem`.