Implement snapshot tests (#241)
This implements snapshot testing, which writes all chunks to the filesystem. Hopefully these'll be a little quicker to test than the evaluation tests, which require us to spin up a browser to execute. The snapshots will detect if a file is expected (in the output dir) but it's not output by turbopack, if a file is output but was not expected, or if the contents has changed.

It also supports a `TURBOPACK_SNAPSHOT_UPDATE=1` env, which will pass all tests, update all files to the latest outputs, and delete any files that shouldn't exist.

The test structure follows a `*/*/input` directory structure. Eg,
```
large-feature-group
└── explicit-feature
    ├── input
    │   └── index.js
    │   └── any-other-files...
    │   └── node_modules/...
    └── output
        └── snapshot_large_feature_group_explicit_feature_input_index.js
        └── snapshot_large_feature_group_explicit_feature_input_node_modules_foo_index.js.js
```

The `*/*/output` is the output chunks we expect turbopack to generate.


A few things are still left todo:
- Diffing of the files isn't supported, which would be very helpful for finding changes.
- I didn't handle static assets, and CSS could be broken.
- It only supports `index.js` as the entry point, but we should support many entrypoints (and different filenames and extensions) via an `options.json`.
- We only perform Browser transformations, but we should support any via an `options.json`.
- I don't know if a nested directory structure is possible for output chunks, so it only supports `output/*` files.