Adding a lockfile interface to hide each package manager's differences
At the moment this interface only deals with data operations and no IO/marshalling. I went with an interface instead of a single struct with an instant for each impl as the contents of a lockfile differ greatly between package managers.

Tested pruning a yarn mono repo made with `create-turbo` and verified the lockfile produced was the same as before this commit.