fix: use docker cache best practices in examples (#1673)
The whole purpose of the `turbo prune --docker` is to use docker caching best practices while the Docker examples do not follow these bests practices.   

Plus, I believe that the first stage of the Dockerfile is wrong as it copies the whole project and defeats the entire purpose of the caching because the first layer never gets cached. I believe the most optimized solution is to run the `turbo prune` before the `docker build` on the machine that runs the `docker build` and set the `out` directory as the docker build context.  

I understand that showing this approach may not be possible in the example, so I suggest we add a comment to the first stage with the above explanation.