Replace some 2 letter variables for readability
While working on another task, I found several things hard to read because of
short variable names. Although types and "Go to Definition" helps in these
cases, long function blocks and large files make it dfficult to focus on the task at hand.
For example, in some cases, the variable `ht` and `th` appear side by
side in the same block!

This commit replaces some of these names to improve general readability
(and also to reduce the diff in future behavior-changing work).