Enable inputs to be relative again. (#1937)
As reported by @dglsparsons, after #1699 `inputs` cannot be relative from the package directory (while still remaining within the project). This issue was published in `turbo@1.4.4`.

This PR patches the traversal behavior so that `inputs` works correctly and adds a test for the behavior to ensure that it does not regress.