fix pruning of importers
We extend the lockfile abstraction to now take a list of workspace package paths in addition to a list of packages.
This is slightly leaky as pnpm is the only package manager where this is strictly necessary although it will help when implementing prune for npm.

Without the explicit information about which workspace dependencies to include we are unable to differentiate between workspace projects where one project has a subset of dependencies of the other.