Support HMR Update Source Maps (#371)
This allows us to reference the source maps for individual chunk items when delivering them for HMR updates.

We currently send down the update via a `JSON.stringify()` of the updated module's code. We then `eval` this code to generate the new "module", and call registered hot handlers with the new module to update their scopes. Thankfully, we can include a `sourceMappingUrl` comment in the evaled code, so we can get away with deferring creating the source map's content until dev tools is opened.

This is a first pass, it really needs to be cleaned up. But at least it's working.


Co-authored-by: Tobias Koppers <1365881+sokra@users.noreply.github.com>