Do not treat dynamically imported chunks as evaluated (#426)
I didn't really understand this part of the code when I wrote this. The dynamic import is never an entrypoint, so we shouldn't treat it as evaluated. Note, the underlying chunk foo.js without the runtime code was always available. But, the client linked to the evaluated foo.js.eval.js, because I incorrectly marked the chunk group as an evaluated one.

Fixes #425