Re-add log message about remote caching (#2122)
This re-adds the log message for Remote Caching, but only when tasks are
actually going to execute. The message was removed in d48fcdf
because it was being included in `--dry-run` output, making
`--dry-run=json` output non-machine-readable. By excluding this log
message from dry run output, we solve both problems.

This commit also moves `logWarning()` into CmdBase to match `logError()`.