Consolidate engine, taskGraph, and scheduler naming to Engine (#2256)
This commit consolidates some internal terminology around the task graph.
"Scheduler", "Engine" and "TaskGraph" are all the same thing. The receiver
on the methods for the now Engine struct was also confusingly named p,
so I've replaced that with e.

This commit also fixes lint warnings/errors in engine.go, since changing the file
caused it to get caught by the linter.