Bugfixes regarding task scope handling (#159)
There are a bunch of missing or incorrect handling of negative counts
that can happen due to parallel processing of adding and removing to
scopes. Negative counts should behave like 0, but the next add will not
reach a positive value yet. This is important if the remove operation
overtakes the add operation.