feature(rust): `bin` command porting (#2234)
Ported `bin` command to Rust. Also ported the help flag to Rust. NOTE: This does mean that if we change commands on the Go side, we will need to fix the Rust help screen. We can disable the Rust help flag and keep the Go one if that is too confusing.

Co-authored-by: Greg Soltis <418301+gsoltis@users.noreply.github.com>