fix: prune should omit descriptors from peerDependencies (#2269)
Partially addresses #2049 

Yarn 2+ doesn't expect descriptors that appear in `peerDependencies` to be kept. This behavior only manifested in a lockfile where a pruned package had a normal dependency descriptor that also appeared as a peer dependency descriptor in a package that wasn't pruned. As background "descriptor" is the term Yarn uses for a semver range that appears in a `package.json` or lockfile entry.

Amended the minimal berry lockfile to now test that we don't keep around descriptors even if they appear in `peerDependencies`.