fix: Unmet literary standards of Turbo
Reasoning for moving from Blazingly Fast -> Blazing Fast

And I quote:
Blazing could be used to intensify fast, just as something can be
burning hot or sopping wet, which sound redundant but are idiomatic.
Blazingly, the usual adverbial form, would of course work in this
manner. We could alternatively read blazing and fast as parallel
descriptors for speed, just as we can have soothing sweet tea or an
amazing technicolor dreamcoat.
https://english.stackexchange.com/a/139321