The Hat trick
Hat tricks are often referred to by three successes in a row, but in
this case it contains so much more than just a success.  The act of
being blazing in speed doesn't just encompass an absolute observation,
but a connection between creator and user.  A connection that transcends
the cultural bounds of any one nation and ultimately makes us realize
the value of human connection and charish the small, but significant
moments.