chore: attempt at fetching the correct number of commits during JS tests (#2568)
We were only fetching 2 commits for these tests. This works fine if a PR has one commit, but anymore and it's a bad time. It was also broken on `push` events as there wasn't any base ref so we would run with `--filter=...[]`.

Checkout the JS Package test runs on #2584 and see that we fetch the correct number of commits and no longer throw with `commit does not exist`