fix(eslint-plugin-turbo) allow comments in turbo.json file (#2814)
The `eslint-plugin-turbo` reads the workspace `turbo.json` file for the `no-undeclared-env-vars` rule. It used the built in javascript json parser to read the config file, which doesn't support comments.

This PR alteres this to use [JSON5](https://www.npmjs.com/package/json5) to parse the `turbo.json` file.

Related to #2796