fix: unable to set verbosity via number (#2993)
Previously users were able to set the verbosity via count (`-vvv`) or by
providing a value (`--verbosity=3`), this PR gives users the ability to
do the latter.

We achieve this by creating our own `Verbosity` struct that when
serialized will be converted to a `u8` allowing us to not change the Go
code. I did remove `-v...` from the help text as it rendered in a
confusing way.

Note:
This doesn't completely match the old behavior as we allowed for funky
stuff like `-v --verbosity=0 --verbosity=3 --verbosity`. (Any guesses
what the verbosity level was set to?)