feat(mdx): initial turbopack-mdx asset (#2979)
Implements / fix WEB-86. This PR is an initial attempt to support mdx
natively inside of turbopack.

It uses mdxrs (https://github.com/wooorm/mdxjs-rs) internally to interop
- compile - mdx inputs into renderable ecma outputs. PR creates a new
type of asset `MDXAsset`, and then let its extension (`.mdx`) can be
consumed by turbopack.

Since mdx is a variant of ecma asset to be rendered, it requires to
perform all of the ecma transforms in the current running turbopack
session. But since its AST is not compatible to plain swc (or ecma), it
also cannot be passed into existing ecmaassets. MDXAsset does interop
for those, by creating a virtualasset for the ecma once mdx is compiled
into plain ecma. This might not be feasible appoarch and requires
different way to chain asset types, something I'd like to address in
this PR. I guess this will be needed for other types of assets in a long
run - where input is not an ecma but requires ecma transforms (i.e vue,
ng templates).