feat(turbo-updater): notifications per channel (#3028)
We can overload name because update-informer allows name in an
owner/name format and then splits the name but includes the owner in the
cached file. So we can do something like canary/turbo, latest/turbo etc.

We should still send a fix upstream to support this more formally. But
this should work for now.

Updated version of https://github.com/vercel/turbo/pull/3017