fix: normalize current directory to non-UNC path on Windows (#3031)
The Windows kernel normalizes paths by converting them to UNC, but silly
things like starting a process with a working directory given as a UNC
path don't work on Windows:
```
'\\?\C:\Users\Chris\Documents\code\junk\yarn-test'
CMD.EXE was started with the above path as the current directory.
UNC paths are not supported.  Defaulting to Windows directory.
```

This PR uses `dunce` when normalizing paths on windows so we produce
non-UNC paths which are better supported by the Windows ecosystem.