Remove chunk item sourcemaps from the asset graph
In order to serve sourcemaps with HMR updates, we (used to) reference a source map asset for every chunk item using the items version hash (see #2464 and #2474 for the headaches this can cause).

This update changes that, so that a chunk becomes responsible for giving access to one of its item's source map. We already have the `GenerateSourceMap` trait which assembles the chunk's `Sectioned` sourcemap from all chunk items, so why not just allow access to an individual section? Coupled with a `ContentSource` that can retrieve the chunk, we can then provide access to that chunk's current item sourcemaps per request.