Fix command to delete local turborepo cache (#3109)
It was in a fenced code block on a single line, which was causing `sh`
to be present at the beginning of each command (very confusing and hard
to catch).

_On a completely unrelated topic..._ (content warning: sarcasm)

_I used [StackBlitz Codeflow](https://stackblitz.com/codeflow/) to make
this PR, and `pnpm run docs` did not work because `make` isn't available
(Codeflow uses a technology called
[WebContainers](https://blog.stackblitz.com/posts/introducing-webcontainers/),
which runs entirely in the browser). I think we should port Make, the
Rust compiler, Cargo and anything needed to compile Turbo to Web
Assembly and/or JavaScript, so that we can run the whole toolchain in
the browser. Sounds pretty easy honestly 👍_

Co-authored-by: Thomas Knickman <tom.knickman@vercel.com>